refactor(frontend): remove the chat journal UI + dead client code
journal-resonance-19: the resonance surfaces (shelf, writing surface, margin
notes, essay modal) fully replace the chat journal, so retire it. Net deletion
(17 files, ~-4400 lines — the chat screen + SSE stream plumbing were large).

Removed:
- Components + their tests: JournalScreen, MessageBubble, ChatInput, TagFilter,
  WeeklyPromptBanner, and Journal.styles.ts. The chat-only composer/message-list
  hooks lived inline in JournalScreen and went with it.
- Client (api/index.ts): botmason.chat + botmason.chatStream, the
  ChatRequest/ChatResponse/ChatStreamCallbacks types, StreamingUnsupportedError,
  and all now-dead SSE plumbing (stream retry/open/read, SSE frame parsing). The
  chat/chat-stream test files (botmason-byok, botmason-stream) went too.
- SearchBar's styles were relocated into its own StyleSheet so Journal.styles.ts
  could be deleted (the new surfaces use JournalEntry/JournalShelf styles).

Kept (verified by grep): useDerivedCurrentWeek (shelf prompt card),
botmason.getBalance/getUsage/addBalance (wallet — resonance charges against it),
the full prompts client. The BYOK key getter now feeds resonance.generate/essay
via byokHeaders, so bring-your-own-key survives the chat removal. No code
references any removed symbol.

Closes #622
Refs #603
