--- frontend/src/features/Journal/ChatInput.tsx
@@ -1,147 +1,0 @@
-import React, { useCallback, useState } from 'react';
-import { Text, TextInput, TouchableOpacity, View } from 'react-native';
-
-import type { JournalTag } from '../../api';
-import { colors } from '../../design/tokens';
-
-import styles from './Journal.styles';
-
-// BUG-FE-JOURNAL-101: server caps the message at 10_000 chars
-// (``JOURNAL_MESSAGE_MAX_LENGTH`` in ``schemas/journal.py``).  Mirroring
-// the cap on the input prevents the user from typing a 50k-char
-// reflection that would surface as a 422 only after Send.
-export const JOURNAL_MESSAGE_MAX_LENGTH = 10_000;
-
-const TAG_OPTIONS: Array<{ value: JournalTag; label: string }> = [
-  { value: 'stage_reflection', label: 'Reflection' },
-  { value: 'practice_note', label: 'Practice' },
-  { value: 'habit_note', label: 'Habit' },
-];
-
-interface ChatInputProps {
-  onSend: (_text: string, _tag?: JournalTag) => void;
-  disabled?: boolean;
-  initialTag?: JournalTag;
-}
-
-interface TagPickerProps {
-  activeTag: JournalTag | undefined;
-  onSelectTag: (_tag: JournalTag | undefined) => void;
-}
-
-const TagPicker = ({ activeTag, onSelectTag }: TagPickerProps): React.JSX.Element => (
-  <View style={styles.tagPickerContainer} testID="tag-picker">
-    {TAG_OPTIONS.map(({ value, label }) => {
-      const isActive = activeTag === value;
-      return (
-        <TouchableOpacity
-          key={value}
-          testID={`tag-option-${value}`}
-          style={[styles.tagPickerOption, isActive && styles.tagPickerOptionActive]}
-          onPress={() => onSelectTag(isActive ? undefined : value)}
-          accessibilityLabel={`Tag as ${label}`}
-          accessibilityRole="button"
-          accessibilityState={{ selected: isActive }}
-        >
-          <Text style={[styles.tagPickerText, isActive && styles.tagPickerTextActive]}>
-            {label}
-          </Text>
-        </TouchableOpacity>
-      );
-    })}
-  </View>
-);
-
-interface InputRowProps {
-  text: string;
-  onChangeText: (_text: string) => void;
-  disabled: boolean;
-  canSend: boolean;
-  onSend: () => void;
-  hasActiveTag: boolean;
-  onToggleTagPicker: () => void;
-}
-
-const InputRow = ({
-  text,
-  onChangeText,
-  disabled,
-  canSend,
-  onSend,
-  hasActiveTag,
-  onToggleTagPicker,
-}: InputRowProps): React.JSX.Element => (
-  <View style={styles.inputContainer}>
-    <TextInput
-      testID="chat-input"
-      accessibilityLabel="Journal reflection input"
-      style={styles.textInput}
-      value={text}
-      onChangeText={onChangeText}
-      placeholder="Write a reflection..."
-      placeholderTextColor={colors.text.tertiary}
-      multiline
-      editable={!disabled}
-      maxLength={JOURNAL_MESSAGE_MAX_LENGTH}
-    />
-    <TouchableOpacity
-      testID="tag-toggle"
-      style={[styles.tagToggleButton, hasActiveTag && styles.tagToggleButtonActive]}
-      onPress={onToggleTagPicker}
-      accessibilityLabel="Toggle tag picker"
-      accessibilityRole="button"
-      accessibilityState={{ expanded: hasActiveTag }}
-    >
-      <Text style={styles.tagToggleText}>#</Text>
-    </TouchableOpacity>
-    <TouchableOpacity
-      testID="send-button"
-      style={[styles.sendButton, !canSend && styles.sendButtonDisabled]}
-      onPress={onSend}
-      disabled={!canSend}
-      accessibilityLabel="Send message"
-      accessibilityRole="button"
-      accessibilityState={{ disabled: !canSend }}
-    >
-      <Text style={styles.sendButtonText}>{'>'}</Text>
-    </TouchableOpacity>
-  </View>
-);
-
-const ChatInput = ({ onSend, disabled = false, initialTag }: ChatInputProps): React.JSX.Element => {
-  const [text, setText] = useState('');
-  const [selectedTag, setSelectedTag] = useState<JournalTag | undefined>(initialTag);
-  const [showTagPicker, setShowTagPicker] = useState(false);
-
-  const handleSend = useCallback(() => {
-    const trimmed = text.trim();
-    if (trimmed.length === 0) return;
-    onSend(trimmed, selectedTag);
-    setText('');
-    setSelectedTag(initialTag);
-    setShowTagPicker(false);
-  }, [text, onSend, selectedTag, initialTag]);
-
-  const toggleTagPicker = useCallback(() => {
-    setShowTagPicker((prev) => !prev);
-  }, []);
-
-  const canSend = text.trim().length > 0 && !disabled;
-
-  return (
-    <View>
-      {showTagPicker && <TagPicker activeTag={selectedTag} onSelectTag={setSelectedTag} />}
-      <InputRow
-        text={text}
-        onChangeText={setText}
-        disabled={disabled}
-        canSend={canSend}
-        onSend={handleSend}
-        hasActiveTag={selectedTag !== undefined}
-        onToggleTagPicker={toggleTagPicker}
-      />
-    </View>
-  );
-};
-
-export default ChatInput;
